오류 수정

+++ src/main/java/com/team1/stelling/domain/vo/SupportVO.java
@@ -16,7 +16,7 @@ public class SupportVO {
    private Long supportSponser; // FK (userNum) 후원한 사람
    private Long subNovelNumber; //FK, 회차번호
    private Long novelNumber; /*FK, 소설 번호*/
-   private Long supportCoin; /*후원 금액 = 차감 코인*/
+   private int supportCoin; /*후원 금액 = 차감 코인*/
    private Date supportDateTime; /*후원 날짜*/
